Move the JobCompletionData class into SimulationResults #6
The JobCompletionData class was only used by SimulationResults, so it was moved into SimulationResults as an inner class. We are still working on a potential solution to eliminate the JobCompletionData class and have it simply be apart of SimulationResults. If this isn't a feasible solution, we will just leave it as an inner class.