Make JobCompletionData a Map with the key (job number) and an ArrayList that contains the completionTime and totalWaitTime #6
This change makes it so the JobCompletionData class doesn't have to be an actual class. We made JobCompletionData into a map that contains the job number as the key and the completionTime and totalWaitTime are now in an ArrayList that serves as the values in the map. The associated tests in SimulationProperties also had to be changed as a result of this change.