Change the ArrayList inside jobCompletions map to an ImmutablePair
JobCompletions[] was an array of Maps with an Integer as the key and an ArrayList to hold the two values. To avoid potential frustration in the future due the mutability of the ArrayList, it was changed to an ImmutablePair. This makes the map a little more readable and also less mutable. All tests pass.